requirements.txt
pyautogui
Pillow
pytesseract
opencv-python
numpy